+++ src/components/BowlingGame/GameControls/GameControl.js
