+++ src/components/BowlingGame/GameControls/GameControl.scss
